Added users and customers apis, modified videos and genre apis

+++ routes/genre.js
@@ -1,4 +1,3 @@
-const debugRoute = require("debug")("vidly:route-genre");
 const debugDb = require("debug")("vidly:db-genre");
 const debug = require("debug")("vidly:log");
 
@@ -10,12 +9,12 @@ const { Genre, validateGenre } = require("../models/genre.js");
 
 router.get("/", (req, res) => {
 
-    Genre.find().sort("updatedOn").select("name _id")
+    Genre.find().sort("-updatedOn").select("name _id")
          .then(r => {
             return res.send(r);
          })
          .catch(e => {
-            debugRoute("Unable to find genres...", e.message);
+            debugDb("Unable to find genres...", e.message);
             res.status(500).send("Internal server error. Please try after sometime.");
          });
 });
@@ -25,7 +24,7 @@ router.post("/", (req, res) => {
     const { error } = validateGenre(req.body);
     
     if(error) return res.status(400).send(error.details[0].message);
-     
+    
     const payload = _.pick(req.body, ["name"]);   
     new Genre(payload).save()
         .then(r => {
@@ -41,12 +40,13 @@ router.post("/", (req, res) => {
 
 router.put("/:id", (req, res) => {
 
+    if( !mongoose.Types.ObjectId.isValid(req.params.id) ) return res.status(400).send("No such customer.");
     
     const { error } = validateGenre(req.body, true);
     if(error) return res.status(400).send(error.details[0].message);
 
     const payload = _.pick(req.body, ["name"]);
-    Genre.findByIdAndUpdate(req.params.id, { $set: payload, $currentDate: {updatedOn: true}}, {new: true, upsert: true})
+    Genre.findByIdAndUpdate(req.params.id, { $set: payload, $currentDate: {updatedOn: true}}, {new: true, upsert: false})
          .then(r => {
              debug("Updating genre...", r);
              if(!r) return res.status(404).send("No such genre.");
@@ -63,11 +63,7 @@ router.put("/:id", (req, res) => {
 
 router.delete("/:id", (req, res) => {
     
-    try {
-        let temp = mongoose.Types.ObjectId(req.params.id);
-    } catch(e) {
-        return res.status(400).send("Not a valid genre id.")
-    };
+    if( !mongoose.Types.ObjectId.isValid(req.params.id) ) return res.status(404).send("No such genre.");
     
     Genre.findByIdAndRemove(req.params.id)
          .then(r => {
@@ -76,7 +72,7 @@ router.delete("/:id", (req, res) => {
              return res.send(r);
          })
          .catch( e => { 
-             debug("Error deleting genre...",e.message); 
+             debugDb("Error deleting genre...",e.message); 
              res.status(500).send("Unable to delete genre. Please try after sometime.");
          });
 });
